TimeSeriesUtils.ts — Function

ts(data; frequency = 1, deltat = 1)

Creates a time series object in R with a certain frequency and observed deltat time steps a part

TimeSeriesUtils.stl — Function

stl(x; frequency = 1, swindow="periodic", twindow = nothing, robust=true)

R function that decompose a time series into seasonal, trend and irregular components using loess.

Examples:

julia> T = 10*12; 
julia> x = 0.01*(1:T) .+ sin.((1:T)*(2pi/12)) .+ 0.2*randn(T)
julia> stl_object = stl(x, frequency = 12)
julia> stl_object

TimeSeriesUtils.mstl — Function

mstl(x; seasonal_periods, swindow = 7 .+ 4 .*(1:6))

R function that decompose a time series into seasonal, trend and irregular components using loess. Decompose a time series into seasonal, trend and remainder components. Allows for multiple seasonal components with different periods.

Examples:

julia> T = 10*12; 
julia> x = 0.01*(1:T) .+ sin.((1:T)*(2pi/12)) .+ sin.((1:T)*(2pi/365.25)) .+ 0.2*randn(T)
julia> mstl_object = mstl(x; seasonal_periods = [12,365.25])
julia> mstl_object

TimeSeriesUtils.nainterpret! — Function

nainterpret!(x; frequency = 1)

R function that interpolates missing values in a time series inplace.

See also nainterpret for an not in place version.

Examples:

julia> nainterpret!([4, missing, 10])
3-element Vector{Float64}:
  4.0
  7.0
 10.0

TimeSeriesUtils.nainterpret — Function

nainterpret(x; frequency = 1)

R function that interpolates missing values in a time series.

See also nainterpret! for an inplace version.

Examples:

julia> y = nainterpret([4, missing, 10])
3-element Vector{Float64}:
  4.0
  7.0
 10.0